move checker to proper place
