+++ dm/pkg/checker/mysql_server_test.go
@@ -14,11 +14,7 @@
 package checker
 
 import (
-	"context"
-
-	"github.com/DATA-DOG/go-sqlmock"
 	tc "github.com/pingcap/check"
-	"github.com/pingcap/tidb-tools/pkg/dbutil"
 )
 
 func (t *testCheckSuite) TestMysqlVersion(c *tc.C) {
@@ -46,95 +42,5 @@ func (t *testCheckSuite) TestMysqlVersion(c *tc.C) {
 		}
 		versionChecker.checkVersion(cs.rawVersion, result)
 		c.Assert(result.State == StateSuccess, tc.Equals, cs.pass)
-	}
-}
-
-func (t *testCheckSuite) TestBinlogDB(c *tc.C) {
-	db, mock, err := sqlmock.New()
-	c.Assert(err, tc.IsNil)
-	ctx := context.Background()
-
-	cases := []struct {
-		flavor   string
-		doDB     string
-		ignoreDB string
-		schemas  map[string]struct{}
-		state    State
-	}{
-		// doDB
-		{
-			doDB: "do",
-			schemas: map[string]struct{}{
-				"do": {},
-			},
-			state: StateSuccess,
-		},
-		{
-			doDB: "do",
-			schemas: map[string]struct{}{
-				"do":  {},
-				"do2": {},
-			},
-			state: StateFailure,
-		},
-		{
-			doDB: "do",
-			schemas: map[string]struct{}{
-				"do2": {},
-			},
-			state: StateFailure,
-		},
-		{
-			doDB: "do, do2",
-			schemas: map[string]struct{}{
-				"do2": {},
-			},
-			state: StateSuccess,
-		},
-		// ignoreDB
-		{
-			ignoreDB: "ignore",
-			schemas: map[string]struct{}{
-				"do":     {},
-				"ignore": {},
-			},
-			state: StateFailure,
-		},
-		{
-			ignoreDB: "ignore",
-			schemas: map[string]struct{}{
-				"do":  {},
-				"do2": {},
-			},
-			state: StateSuccess,
-		},
-		{
-			ignoreDB: "ignore, ignore2",
-			schemas: map[string]struct{}{
-				"do":      {},
-				"ignore2": {},
-			},
-			state: StateFailure,
-		},
-		{
-			ignoreDB: "ignore, ignore2",
-			schemas: map[string]struct{}{
-				"ignore3": {},
-			},
-			state: StateSuccess,
-		},
-	}
-
-	for _, cs := range cases {
-		binlogDBChecker := NewBinlogDBChecker(db, &dbutil.DBConfig{}, cs.schemas)
-		versionRow := sqlmock.NewRows([]string{"Variable_name", "Value"}).AddRow("version", "mysql")
-		masterStatusRow := sqlmock.NewRows([]string{"File", "Position", "Binlog_Do_DB", "Binlog_Ignore_DB", "Executed_Gtid_Set"}).
-			AddRow("", 0, cs.doDB, cs.ignoreDB, "")
-		mock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(versionRow)
-		mock.ExpectQuery("SHOW MASTER STATUS").WillReturnRows(masterStatusRow)
-
-		r := binlogDBChecker.Check(ctx)
-		c.Assert(mock.ExpectationsWereMet(), tc.IsNil)
-		c.Assert(r.State, tc.Equals, cs.state)
 	}
 }
